Move everything inside NistPubid module
