+++ lib/nist_pubid/version.rb
@@ -1,0 +1,3 @@
+module NistPubid
+  VERSION = "0.1.1".freeze
+end
